+++ AU/Public/Set-DescriptionFromReadme.ps1
@@ -5,31 +5,29 @@
 .DESCRIPTION
   This script should be called in au_AfterUpdate to put the text in the README.md
   into description tag of the Nuspec file. The current description will be replaced.
-  Function will throw an error if README.md is not found.
-
-.PARAMETER SkipFirst
-  Number of start lines to skip from the README.md, by default 0.
-
-
-.PARAMETER SkipLast
-  Number of end lines to skip from the README.md, by default 0.
+  
+  You need to call this function manually only if you want to pass it custom parameters.
+  In that case use NoReadme parameter of the Update-Package.
 
 .EXAMPLE
-  function global:au_AfterUpdate  { Set-DescriptionFromReadme -SkipFirst 2 }
+  function global:au_AfterUpdate  { Set-DescriptionFromReadme -Package $args[0] -SkipLast 2 -SkipFirst 2 }
 #>
-function Set-DescriptionFromReadme([int]$SkipFirst=0, [int]$SkipLast=0) {
-    if (!(Test-Path README.md)) { throw 'Set-DescriptionFromReadme: README.md not found' }
+function Set-DescriptionFromReadme{
+    param(
+      [AUPackage] $Package, 
+      # Number of start lines to skip from the README.md, by default 0.
+      [int] $SkipFirst=0, 
+      # Number of end lines to skip from the README.md, by default 0.
+      [int] $SkipLast=0
+    )
 
-    Write-Host 'Setting README.md to Nuspec description tag'
+    'Setting README.md to Nuspec description tag'
+
     $description = gc README.md -Encoding UTF8
     $endIdx = $description.Length - $SkipLast
     $description = $description | select -Index ($SkipFirst..$endIdx) | Out-String
+    $description = "<![CDATA[" + $description + "]]>"
 
-    $nuspecFileName = $Latest.PackageName + ".nuspec"
-    $nu = gc $nuspecFileName -Raw -Encoding UTF8
-    $nu = $nu -replace "(?smi)(\<description\>).*?(\</description\>)", "`${1}$($description)`$2"
-
-    $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding($False)
-    $NuPath = (Resolve-Path $NuspecFileName)
-    [System.IO.File]::WriteAllText($NuPath, $nu, $Utf8NoBomEncoding)
+    $Package.NuspecXml.package.metadata.description = $description
+    $Package.SaveNuspec()
 }